backend/quiz_game_table/fetchcategories/requirements.txt
Flask==2.0.1
google-cloud-firestore==2.3.2
google-cloud-language==2.4.0
